Refacter: Using .tsx. Have bug so i commend the code in ItemScreen.tsx in line 40 & ItemCarContainer.tsx in line 18
