+++ .gitignore
@@ -33,5 +33,3 @@ yarn-error.*
 
 # typescript
 *.tsbuildinfo
-
-yarn.lock
