+++ App.tsx
@@ -2,7 +2,7 @@ import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Tabs from './navigation/tabs.js';
+import Tabs from './navigation/tabs';
 
 const Stack = createNativeStackNavigator();
