+++ README.md
@@ -1,0 +1,7 @@
+# NINH BINH HERITAGE
+## Technology
+- _React Native with Expo_
+- _React Navigation_: Navigation between screens
+- _React Navigation Stack_: Để di chuyển giữa các màn hình, lưu trữ lịch sử điều hướng để có thể goBack()
+- _React Native Safe Area Context_: Vùng safe area là vùng tránh tai thỏ
+- _React Navigation Bottom Tabs_: Để tạo bottom tab
