+++ components/ItemCarContainer.tsx
@@ -2,13 +2,21 @@ import { View, Text, TouchableOpacity, Image } from "react-native";
 import { FontAwesome } from "@expo/vector-icons";
 import React from "react";
 import { useNavigation } from "@react-navigation/native";
+import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 
-const ItemCarDontainer = ({ imageSrc, title, location, data }) => {
+type Props = {
+    imageSrc: string;
+    title?: string;
+    location: string;
+    data: string;
+};
+
+const ItemCarContainer = ({ imageSrc, title, location, data }: Props) => {
     const navigation = useNavigation();
     return (
         <TouchableOpacity
-            onPress={() => navigation.navigate("ItemScreen", { param: data })}
-            className="rounded-md border border-gray-300 space-y-2 px-3 py-2 shadow-md bg-white w-[182px] my-2"
+            // onPress={() => navigation.navigate("ItemScreen", { param: data })}
+            className="rounded-md border border-gray-300 space-y-2 px-3 py-2 shadow-md bg-white     w-[182px] my-2"
         >
             <Image
                 source={{ uri: imageSrc }}
@@ -24,7 +32,7 @@ const ItemCarDontainer = ({ imageSrc, title, location, data }) => {
                     <View className="flex-row items-center space-x-1">
                         <FontAwesome name="map-marker" size={20} color="#8597A2" />
                         <Text className="text-[#428288] text-[14px] font-bold">
-                            {location?.length > 18 ? `${title.slice(0, 18)}..` : location}
+                            {location.length > 18 ? `${title.slice(0, 18)}..` : location}
                         </Text>
                     </View>
                 </>
@@ -35,4 +43,4 @@ const ItemCarDontainer = ({ imageSrc, title, location, data }) => {
     );
 };
 
-export default ItemCarDontainer;
+export default ItemCarContainer;
