+++ components/MenuContainer.tsx
@@ -1,7 +1,14 @@
 import { View, Text, TouchableOpacity, Image } from "react-native";
 import React from "react";
 
-const MenuContainer = ({ title, imageSrc, type, setType }) => {
+type Props = {
+    title: string;
+    imageSrc: string;
+    type: string;
+    setType: any;
+};
+
+const MenuContainer = ({ title, imageSrc, type, setType }: Props) => {
     const handlePress = () => {
         setType(title.toLowerCase());
     };
@@ -14,7 +21,7 @@ const MenuContainer = ({ title, imageSrc, type, setType }) => {
                 className={`w-24 h-24 p-2 shadow-sm rounded-full items-center justify-center ${type === title.toLowerCase() ? "bg-gray-200" : ""
                     }`}
             >
-                <Image source={imageSrc} className="w-full h-full object-contain" />
+                <Image source={{ uri: imageSrc }} className="w-full h-full object-contain" />
             </View>
             <Text className="text-[#00BCC9] text-xl font-semibold">{title}</Text>
         </TouchableOpacity>
