+++ package.json
@@ -27,7 +27,9 @@
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.14",
-    "tailwindcss": "^3.3.2"
+    "@types/react-native": "^0.72.7",
+    "tailwindcss": "^3.3.2",
+    "typescript": "^5.1.3"
   },
   "private": true
 }
