+++ screens/Attraction.tsx
@@ -1,7 +1,7 @@
 import { SafeAreaView } from "react-native";
 import { Text, View } from 'react-native';
 
-export default AttractionScreen = () => {
+const AttractionScreen = () => {
   return (
     <SafeAreaView>
       <View className="flex-row items-center justify-between px-8">
@@ -13,3 +13,5 @@ export default AttractionScreen = () => {
     </SafeAreaView>
   )
 }
+
+export default AttractionScreen;
